Switch to builder class as yql builder
