+++ src/cpr_data_access/vespa.py
@@ -63,122 +63,6 @@ def find_vespa_cert_paths() -> tuple[Path, Path]:
     return cert_path, key_path
 
 
-def sanitize(user_input: str) -> str:
-    """
-    Sanitize user input strings
-
-    This is intended to limit possible YQL injection attacks. The query endpoint is not
-    as vulnerable as sql as updates/inserts/deletes in vespa are handled by a seperate
-    endpoint. The main purpose here is to mitigate vespas "INVALID_QUERY_PARAMETER"
-    errors. See vespa codebase for context on full list of errors:
-    https://github.com/vespa-engine/vespa/blob/dd94d619668210d09792597cbd218994058e923e
-    /container-core/src/main/java/com/yahoo/container/protect/Error.java#L15C2-L15C2
-
-    :param str user_input: a potentially hazardous user input string
-    :return str: sanitized user input string
-    """
-    # in the generated YQL string, user inputs are wrapped in double quotes. We should
-    # therefore remove any double quotes from the user inputs to avoid early terminations,
-    # which could allow for subsequent injections
-    user_input = user_input.replace('"', "")
-
-    # remove backslashes, as these are used by vespa as an escape character
-    user_input = user_input.replace("\\", " ")
-
-    # remove any extra whitespace from the user input string
-    user_input = " ".join(user_input.split())
-
-    return user_input
-
-
-def build_yql(request: SearchParameters, sensitive: bool = False) -> str:
-    """
-    Build a YQL string for retrieving relevant, filtered, sorted results from vespa
-
-    :param SearchParameters request: a search request object comprised of the user's
-        search parameters
-    :return str: formatted YQL which incorporates the user's search parameters
-    """
-    request.query_string = sanitize(request.query_string)
-
-    if request.exact_match:
-        rendered_query_string_match = f"""
-            where (
-                (family_name contains({{stem: false}}"{request.query_string}")) or
-                (family_description contains({{stem: false}}"{request.query_string}")) or
-                (text_block contains ({{stem: false}}"{request.query_string}"))
-            )
-        """
-    elif sensitive:
-        rendered_query_string_match = f"""
-            where (
-                {{"targetHits": 1000}} weakAnd(
-                    family_name contains "{ request.query_string }",
-                    family_description contains "{ request.query_string }",
-                    text_block contains "{ request.query_string }"
-            ))
-        """
-    else:
-        rendered_query_string_match = f"""
-            where ((
-                {{"targetHits": 1000}} weakAnd(
-                    family_name contains "{ request.query_string }",
-                    family_description contains "{ request.query_string }",
-                    text_block contains "{ request.query_string }"
-                )
-            ) or (
-                [{{"targetNumHits": 1000}}]
-                nearestNeighbor(family_description_embedding,query_embedding)
-            ) or (
-                [{{"targetNumHits": 1000}}]
-                nearestNeighbor(text_embedding,query_embedding)
-            ))
-        """
-
-    rendered_filters = ""
-    if request.keyword_filters:
-        filters = []
-        for field_name, values in request.keyword_filters.items():
-            for value in values:
-                filters.append(f'({field_name} contains "{sanitize(value)}")')
-        if filters:
-            rendered_filters = f" and ({' or '.join(filters)})"
-
-    if request.year_range:
-        start, end = request.year_range
-        if start:
-            rendered_filters += f" and (family_publication_year >= {start})"
-        if end:
-            rendered_filters += f" and (family_publication_year <= {end})"
-
-    rendered_continuation = (
-        f"{{ 'continuations':['{request.continuation_token}'] }}"
-        if request.continuation_token
-        else ""
-    )
-
-    rendered_query = f"""
-        select *
-        from sources family_document, document_passage
-        { rendered_query_string_match }
-        { rendered_filters }
-        limit 0
-        |
-        { rendered_continuation }
-        all(
-            group(family_import_id)
-            max({request.limit})
-            each(
-                max({request.max_hits_per_family})
-                each(output(summary(search_summary))
-            )
-        )
-    )
-    """
-
-    return " ".join(rendered_query.split())
-
-
 def parse_vespa_response(
     request: SearchParameters,
     vespa_response: VespaResponse,
