+++ src/cpr_data_access/yql_builder.py
@@ -2,7 +2,34 @@ from string import Template
 from typing import Optional
 
 from cpr_data_access.models.search import SearchParameters
-from cpr_data_access.vespa import sanitize
+
+
+def sanitize(user_input: str) -> str:
+    """
+    Sanitize user input strings
+
+    This is intended to limit possible YQL injection attacks. The query endpoint is not
+    as vulnerable as sql as updates/inserts/deletes in vespa are handled by a seperate
+    endpoint. The main purpose here is to mitigate vespas "INVALID_QUERY_PARAMETER"
+    errors. See vespa codebase for context on full list of errors:
+    https://github.com/vespa-engine/vespa/blob/dd94d619668210d09792597cbd218994058e923e
+    /container-core/src/main/java/com/yahoo/container/protect/Error.java#L15C2-L15C2
+
+    :param str user_input: a potentially hazardous user input string
+    :return str: sanitized user input string
+    """
+    # in the generated YQL string, user inputs are wrapped in double quotes. We should
+    # therefore remove any double quotes from the user inputs to avoid early terminations,
+    # which could allow for subsequent injections
+    user_input = user_input.replace('"', "")
+
+    # remove backslashes, as these are used by vespa as an escape character
+    user_input = user_input.replace("\\", " ")
+
+    # remove any extra whitespace from the user input string
+    user_input = " ".join(user_input.split())
+
+    return user_input
 
 
 class YQLBuilder:
